# contracts/internal/internal-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: SuperAgentSystem Internal API
  version: v1
  summary: Launch Spec v1 frozen internal contracts with Memory System V3 single-tenant extension
servers:
  - url: https://internal.super-agent.local
components:
  parameters:
    TenantId:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
    Scope:
      name: X-Scope
      in: header
      required: true
      schema:
        type: string
    TraceId:
      name: X-Trace-Id
      in: header
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
  responses:
    ErrorResponse:
      description: Standard error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error_code, message, trace_id, retryable, details]
      properties:
        error_code:
          type: string
        message:
          type: string
        trace_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
    Step:
      type: object
      required:
        - step_key
        - step_order
        - title
        - step_type
        - dependency_keys
        - risk_level
        - side_effect_class
        - input_payload
        - expected_output
        - acceptance_criteria
      properties:
        step_key:
          type: string
        step_order:
          type: integer
        title:
          type: string
        step_type:
          type: string
        dependency_keys:
          type: array
          items:
            type: string
        risk_level:
          type: string
          enum: [low, medium, high, critical]
        side_effect_class:
          type: string
          enum: [none, read_only, external_resource, state_change, approval]
        capability_hint:
          type: string
        compensation_hint:
          type: object
          additionalProperties: true
        input_payload:
          type: object
          additionalProperties: true
        expected_output:
          type: object
          additionalProperties: true
        acceptance_criteria:
          type: array
          items:
            type: object
            additionalProperties: true
    PlanRequest:
      type: object
      required: [task_request_id, task_type, goal, normalized_envelope, resident_context, retrieval_budget]
      properties:
        task_request_id:
          type: string
          format: uuid
        task_type:
          type: string
        goal:
          type: string
        normalized_envelope:
          type: object
          additionalProperties: true
        resident_context:
          type: object
          additionalProperties: true
        retrieval_budget:
          type: object
          additionalProperties: true
    PlanResponse:
      type: object
      required: [task_plan_id, plan_version, risk_level, acceptance_criteria, steps]
      properties:
        task_plan_id:
          type: string
          format: uuid
        plan_version:
          type: integer
        risk_level:
          type: string
        acceptance_criteria:
          type: array
          items:
            type: object
            additionalProperties: true
        steps:
          type: array
          items:
            $ref: "#/components/schemas/Step"
    ResolveRequest:
      type: object
      required: [task_plan_id, task_step_id, step_type, risk_level, side_effect_class, required_scopes, fingerprint_context]
      properties:
        task_plan_id:
          type: string
          format: uuid
        task_step_id:
          type: string
          format: uuid
        step_type:
          type: string
        risk_level:
          type: string
        side_effect_class:
          type: string
        required_scopes:
          type: array
          items:
            type: string
        fingerprint_context:
          type: object
          additionalProperties: true
    ResolveResponse:
      type: object
      required: [resolved_capability_id, candidate_capabilities, approval_required, resolution_reason]
      properties:
        resolved_capability_id:
          type: string
          format: uuid
        candidate_capabilities:
          type: array
          items:
            type: object
            additionalProperties: true
        approval_required:
          type: boolean
        resolution_reason:
          type: string
    DispatchRequest:
      type: object
      required: [task_request_id, task_plan_id, task_step_id, resolved_capability_id, dispatch_payload, precheck_token]
      properties:
        task_request_id:
          type: string
          format: uuid
        task_plan_id:
          type: string
          format: uuid
        task_step_id:
          type: string
          format: uuid
        resolved_capability_id:
          type: string
          format: uuid
        dispatch_payload:
          type: object
          additionalProperties: true
        precheck_token:
          type: string
    DispatchResponse:
      type: object
      required: [dispatch_status, attempt_no, execution_reference, journal_checkpoint, stream_state]
      properties:
        dispatch_status:
          type: string
        attempt_no:
          type: integer
        execution_reference:
          type: object
          additionalProperties: true
        journal_checkpoint:
          type: string
        stream_state:
          type: string
          enum: [provisional, committed, revoked, replanned, blocked]
...
paths:
  /internal/planner/plan:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanRequest"
      responses:
        "200":
          description: Planned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlanResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/resolver/resolve:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveRequest"
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /internal/router/dispatch:
    post:
      parameters:
        - $ref: "#/components/parameters/TenantId"
        - $ref: "#/components/parameters/Scope"
        - $ref: "#/components/parameters/TraceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DispatchRequest"
      responses:
        "200":
          description: Dispatched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DispatchResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...